# Big ints / long floats optimization