@lemire

The speedups are not related to this change, without them there's around [+2%; +3%] slowdown, but they seem to compensate for the extra added logic. I'm particularly happy that mesh.txt benchmarks (that are super sensitive to instruction count) are not hurt in the end.

We also seem to have made it past lexical in the big_ints case which is nice.

Benchmarks:

### `big_ints`

```
fast_float (before) 	175.31 ns
fast_float (after)	83.38 ns	(-52.4%)
lexical_core     	94.86 ns
from_str        	452.05 ns
```

### `uniform`

```
fast_float (before) 	19.98 ns
fast_float (after)   	20.30 ns	(+1.1%)
lexical_core     	53.85 ns
from_str        	101.32 ns
```

### `canada.txt`

```
fast_float (before)	22.29 ns
fast_float (after)  	22.05 ns	(-1.1%)
lexical_core     	64.61 ns
from_str        	173.28 ns
```

### `mesh.txt`

```
fast_float (before) 	10.70 ns
fast_float (after)  	10.71 ns	(+0.0%)
lexical_core     	22.98 ns
from_str         	22.52 ns
```